+++ src/components/DownloadComponents/helpers/createDownloadResourcesLink.test.tsx
@@ -22,7 +22,8 @@ describe("createDownloadResourcesLink()", () => {
   it("should return correct data if fetch is successful", async () => {
     downloadResourcesLink = await createDownloadResourcesLink(
       "lesson-slug",
-      "exit-quiz-answers,worksheet-pdf"
+      "exit-quiz-answers,worksheet-pdf",
+      "teachers"
     );
 
     expect(downloadResourcesLink).toEqual(data.url);
@@ -36,7 +37,8 @@ describe("createDownloadResourcesLink()", () => {
     try {
       await createDownloadResourcesLink(
         "lesson-slug",
-        "exit-quiz-answers,worksheet-pdf"
+        "exit-quiz-answers,worksheet-pdf",
+        "teachers"
       );
     } catch (error) {
       expect(error).toEqual("bad thing");
@@ -57,7 +59,8 @@ describe("createDownloadResourcesLink()", () => {
     try {
       await createDownloadResourcesLink(
         "lesson-slug",
-        "exit-quiz-answers,worksheet-pdf"
+        "exit-quiz-answers,worksheet-pdf",
+        "teachers"
       );
     } catch (error) {
       expect((error as Error).message).toEqual("specific error");
@@ -78,10 +81,71 @@ describe("createDownloadResourcesLink()", () => {
     try {
       await createDownloadResourcesLink(
         "lesson-slug",
-        "exit-quiz-answers,worksheet-pdf"
+        "exit-quiz-answers,worksheet-pdf",
+        "teachers"
       );
     } catch (error) {
       expect((error as Error).message).toEqual("API error");
+    }
+  });
+  it("should fetch from legacy vercel legacy vercel api if viewType is teachers", async () => {
+    await createDownloadResourcesLink(
+      "lesson-slug",
+      "exit-quiz-answers,worksheet-pdf",
+      "teachers"
+    );
+
+    expect(global.fetch).toBeCalledWith(
+      "https://api.thenational.academy/api/downloads/lesson/lesson-slug?selection=exit-quiz-answers,worksheet-pdf"
+    );
+  });
+  it("should fetch from download api if viewType is teachers-2023", async () => {
+    await createDownloadResourcesLink(
+      "lesson-slug",
+      "exit-quiz-answers,worksheet-pdf",
+      "teachers-2023"
+    );
+
+    expect(global.fetch).toBeCalledWith(
+      "https://downloads-api.thenational.academy/api/lesson/lesson-slug/download?selection=exit-quiz-answers,worksheet-pdf"
+    );
+  });
+  it("should throw an error when NEXT_PUBLIC_DOWNLOAD_API_URL is not defined", async () => {
+    const originalEnv = process.env;
+    delete process.env.NEXT_PUBLIC_DOWNLOAD_API_URL;
+
+    try {
+      await createDownloadResourcesLink(
+        "lesson-slug",
+        "exit-quiz-answers,worksheet-pdf",
+        "teachers-2023"
+      );
+    } catch (error) {
+      expect(error).toEqual(
+        new TypeError(
+          "process.env.NEXT_PUBLIC_DOWNLOAD_API_URL must be defined"
+        )
+      );
+    } finally {
+      process.env = originalEnv;
+    }
+  });
+  it("should throw an error when NEXT_PUBLIC_VERCEL_API_URL is not defined", async () => {
+    const originalEnv = process.env;
+    delete process.env.NEXT_PUBLIC_VERCEL_API_URL;
+
+    try {
+      await createDownloadResourcesLink(
+        "lesson-slug",
+        "exit-quiz-answers,worksheet-pdf",
+        "teachers-2023"
+      );
+    } catch (error) {
+      expect(error).toEqual(
+        new TypeError("process.env.NEXT_PUBLIC_VERCEL_API_URL must be defined")
+      );
+    } finally {
+      process.env = originalEnv;
     }
   });
 });